LayoutTests/inspector/runtime/evaluate-userGestureEmulation-userIsInteracting-expected.txt

Tests for Runtime emulateUserGesture option.


== Running test suite: Runtime.evaluateUserGestureEmulation.userIsInteracting
-- Running test case: EvaluateWithoutEmulatingUserGesture.userIsInteracting
User is NOT Interacting
-- Running test case: EvaluateWithEmulatingUserGesture.userIsInteracting
User is Interacting
